backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
tensorflow==2.15.0
pillow==10.2.0
numpy==1.26.3
pydantic==2.5.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
opencv-python==4.9.0.80
python-dotenv==1.0.0
aiofiles==23.2.1
scikit-learn==1.4.0
scipy==1.12.0
pyyaml==6.0.1
qrcode[pil]==7.4.2
pandas==2.2.0
torch==2.1.2
torchvision==0.16.2
plotly==5.18.0
kaleido==0.2.1
websockets==12.0
